+++ CHANGELOG.md
@@ -6,12 +6,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ⚡️(frontend) export html #1669
+
 ### Changed
 
 - ♿(frontend) improve accessibility:
   - ♿(frontend) add skip to content button for keyboard accessibility #1624
   - ♿(frontend) fix toggle panel button a11y labels #1634
-- ⚡️(frontend) Enhance/html copy to download #1669
+- ⚡️(frontend) improve Comments feature #1687
 
 ### Fixed